Migration from group project
I have a collection of projects in 1 repository; nicknamed Senalis - Star Wars moon of Rutan, which held all my WIP projects. Since this project expanded to be more standalone IE the use of containers I felt it needed it's own repository. It also means it can be public which is nice.