Local debug & GenericHttpClient
Added a PS script to restore, build, publish and launch the client app so that it can be debugged without containers. Services works as is.
Removed the baseAddress param on the GenericHttpClient as it is/was currently only used to link between Client -> Services. It now gets the ApiUrl so doesn't require a constructor parameter.

I would be interested to look at a message bus (RabbitMQ or Grpc) to do cross domain communication. This looks as though it would reduce this level of coupling between the apps. For now HttpClient will suffice.