Testing Client Bridges
Added an interface for the GenericHttpClient :/ I want to test the behaviour in the Client (User facing app) so that the types of objects returned are not amended. For this moving to use an interface seemed like the least invasive idea.

As the tests are added the bridges will be updated to use the interface over the concrete class.

Also began to build logic to aid in tests for the AchievementsBridge.