Task.FromResult, IGenericHttpClient & Removal of unused calls
Removed all instances of Task.FromResult(x).Result from the ClientBridges. In the GenericHttpClient there are already awaiters present and in the instances where we return a response body this is converted in async so these are not required.

Moved all ServicesBridges to use IGenericHttpClient so that behaviour can be easily mocked out.

Removed the unused call to GetAllAchievements from the AchievementsBridge; this didn't appear used.

Moved to use DeserializeAnonymousType from the api responses rather than the DeserialiseObject as the current returned types are anonymous so this matches the pattern better.